+++ reactjs/src/components/CommentFooter.js
@@ -1,13 +1,32 @@
 import React from 'react';
 import AnyCommentComponent from "./AnyCommentComponent";
 import Icon from './Icon'
-import {faReply, faHeart, faEdit, faTrashAlt} from '@fortawesome/free-solid-svg-icons'
+import {faReply, faHeart, faEdit} from '@fortawesome/free-solid-svg-icons'
 
 /**
  * CommentFooter renders single comment actions such as
  * edit, reply, like, etc.
  */
 class CommentFooter extends AnyCommentComponent {
+
+    constructor(props) {
+        super();
+
+        this.state = {
+            processLikeStates: false
+        };
+    }
+
+    /**
+     * Handle like action.
+     */
+    handleLike = (e) => {
+        this.setState({
+            processLikeStates: !this.state.processLikeStates
+        });
+
+        this.props.onLike(e);
+    };
 
     render() {
         const settings = this.props.settings;
@@ -17,7 +36,12 @@ class CommentFooter extends AnyCommentComponent {
         const replyIcon = <Icon icon={faReply}/>;
         const likeIcon = <Icon icon={faHeart} style={this.props.hasLike ? {color: '#EC4568'} : ''}/>;
         const editIcon = <Icon icon={faEdit}/>;
-        const trashIcon = <Icon icon={faTrashAlt}/>;
+
+        let likeActionClass = "anycomment comment-single-body__actions-like" + (this.props.hasLike ? '-active' : '');
+
+        if (this.state.processLikeStates) {
+            likeActionClass += " comment-single-body__actions-like-" + (this.props.hasLike ? 'active' : 'static')
+        }
 
         return (
             <footer className="anycomment comment-single-body__actions">
@@ -27,8 +51,8 @@ class CommentFooter extends AnyCommentComponent {
                     </li>
                     <li className="anycomment">
                         <a href="javascript:void(0)"
-                           className="anycomment comment-single-body__actions-like"
-                           onClick={(e) => this.props.onLike(e)}>{likeIcon}
+                           className={likeActionClass}
+                           onClick={(e) => this.handleLike(e)}>{likeIcon}
                             <span itemProp="upvoteCount">{this.props.likesCount}</span>
                         </a>
                     </li>